requirements.txt
# Core Application & CLI
rich==13.7.1
typer[all]
# Data Handling
pandas==2.2.2
# logspip
structlog

# Remote Execution
paramiko==3.4.0
pywinrm==0.5.0
pyvmomi==8.0.3.0.1
git+https://github.com/vmware/vsphere-automation-sdk-python.git
requests==2.31.0

# Security & Secrets Management
keyring==25.2.1
keyrings.alt==5.0.0 

setuptools<70

# Graph Analysis
networkx==3.3
python-louvain==0.16

# Graph Database Driver
neo4j==5.20.0

# YAML Handling
PyYAML==6.0.1

# NMAP Scanning
python-nmap==0.7.1

# System Information
pywin32==311